+++ allure-report/data/behaviors.csv
@@ -1,2 +1,2 @@
 "BROKEN","EPIC","FAILED","FEATURE","PASSED","SKIPPED","STORY","UNKNOWN"
-"1","","4","","43","1","","0"
+"0","","0","","47","1","","0"
