--- allure-report/data/categories.csv
@@ -1,3 +1,0 @@
-"BROKEN","CATEGORY","FAILED","PASSED","SKIPPED","UNKNOWN"
-"0","Product defects","4","0","0","0"
-"1","Test defects","0","0","0","0"
