+++ allure-report/data/categories.json
@@ -1,1 +1,1 @@
-{"uid":"4b4757e66a1912dae1a509f688f20b0f","name":"categories","children":[{"name":"Test defects","children":[{"name":"Test timeout of 90000ms exceeded.","children":[{"name":"navigate to The Good Guys login page","uid":"dcc46cc25d6cdac9","parentUid":"95ba3cf3ac2f598a5eab414624c5abea","status":"broken","time":{"start":1752973871942,"stop":1752973963864,"duration":91922},"flaky":false,"newFailed":false,"newPassed":false,"newBroken":false,"retriesCount":1,"retriesStatusChange":false,"parameters":[],"tags":[]}],"uid":"95ba3cf3ac2f598a5eab414624c5abea"}],"uid":"bdbf199525818fae7a8651db9eafe741"},{"name":"Product defects","children":[{"name":"Error: Timed out 5000ms waiting for expect(locator).toBeVisible()\n\nLocator: locator('#personal_info_title')\nExpected: visible\nReceived: <element(s) not found>\nCall log:\n  - Expect \"toBeVisible\" with timeout 5000ms\n  - waiting for locator('#personal_info_title')\n","children":[{"name":"GGRegPageDropDown Test","uid":"98ed98dd70985653","parentUid":"a1155472b0f0e0cba2771fa649f2caf1","status":"failed","time":{"start":1752973882824,"stop":1752973889575,"duration":6751},"flaky":false,"newFailed":false,"newPassed":false,"newBroken":false,"retriesCount":1,"retriesStatusChange":false,"parameters":[],"tags":[]}],"uid":"a1155472b0f0e0cba2771fa649f2caf1"},{"name":"Error: apiRequestContext.post: getaddrinfo ENOTFOUND mock-api.example.com\nCall log:\n  - → POST https://mock-api.example.com/login\n    - user-agent: Playwright/1.54.1 (x64; windows 10.0) node/22.11\n    - accept: */*\n    - accept-encoding: gzip,deflate,br\n    - Content-Type: application/json\n    - content-length: 41\n","children":[{"name":"Authenticate and get token (simulate)","uid":"5fcb52be5d386383","parentUid":"43c83fb7930409db46e49310fae69897","status":"failed","time":{"start":1752973926356,"stop":1752973926641,"duration":285},"flaky":false,"newFailed":false,"newPassed":false,"newBroken":false,"retriesCount":1,"retriesStatusChange":false,"parameters":[],"tags":[]}],"uid":"43c83fb7930409db46e49310fae69897"},{"name":"Error: expect(received).not.toBe(expected) // Object.is equality\n\nExpected: not \"\"","children":[{"name":"Use auth token in another API call","uid":"282f2417b3fa088e","parentUid":"ae559acf0229be05913d79bccec7fe56","status":"failed","time":{"start":1752973929259,"stop":1752973929283,"duration":24},"flaky":false,"newFailed":false,"newPassed":false,"newBroken":false,"retriesCount":0,"retriesStatusChange":false,"parameters":[],"tags":[]}],"uid":"ae559acf0229be05913d79bccec7fe56"},{"name":"Error: locator.click: Target page, context or browser has been closed\nCall log:\n  - waiting for getByRole('link', { name: '' })\n","children":[{"name":"test","uid":"41ac158a11feebb1","parentUid":"b611060e02cc21b19c01f19b5e222ce3","status":"failed","time":{"start":1752973935945,"stop":1752973990725,"duration":54780},"flaky":false,"newFailed":false,"newPassed":false,"newBroken":false,"retriesCount":0,"retriesStatusChange":false,"parameters":[],"tags":[]}],"uid":"b611060e02cc21b19c01f19b5e222ce3"}],"uid":"8fb3a91ba5aaf9de24cc8a92edc82b5d"}]}
+{"uid":"4b4757e66a1912dae1a509f688f20b0f","name":"categories","children":[]}
