+++ allure-report/export/influxDbData.txt
@@ -1,15 +1,13 @@
-launch_status failed=4 1752974173000000000
-launch_status broken=1 1752974173000000000
-launch_status passed=43 1752974173000000000
-launch_status skipped=1 1752974173000000000
-launch_status unknown=0 1752974173000000000
-launch_time duration=118783 1752974173000000000
-launch_time min_duration=1 1752974173000000000
-launch_time max_duration=91922 1752974173000000000
-launch_time sum_duration=328536 1752974173000000000
-launch_time start=1752973871942 1752974173000000000
-launch_time stop=1752973990725 1752974173000000000
-launch_problems product_defects=4 1752974173000000000
-launch_problems test_defects=1 1752974173000000000
-launch_retries retries=75 1752974173000000000
-launch_retries run=49 1752974173000000000
+launch_status failed=0 1753964748000000000
+launch_status broken=0 1753964748000000000
+launch_status passed=47 1753964748000000000
+launch_status skipped=1 1753964748000000000
+launch_status unknown=0 1753964748000000000
+launch_time duration=31026034 1753964748000000000
+launch_time min_duration=1 1753964748000000000
+launch_time max_duration=67996 1753964748000000000
+launch_time sum_duration=347115 1753964748000000000
+launch_time start=1753933170729 1753964748000000000
+launch_time stop=1753964196763 1753964748000000000
+launch_retries retries=15 1753964748000000000
+launch_retries run=48 1753964748000000000
