+++ allure-report/export/prometheusData.txt
@@ -1,15 +1,13 @@
-launch_status_failed 4
-launch_status_broken 1
-launch_status_passed 43
+launch_status_failed 0
+launch_status_broken 0
+launch_status_passed 47
 launch_status_skipped 1
 launch_status_unknown 0
-launch_time_duration 118783
+launch_time_duration 31026034
 launch_time_min_duration 1
-launch_time_max_duration 91922
-launch_time_sum_duration 328536
-launch_time_start 1752973871942
-launch_time_stop 1752973990725
-launch_problems_product_defects 4
-launch_problems_test_defects 1
-launch_retries_retries 75
-launch_retries_run 49
+launch_time_max_duration 67996
+launch_time_sum_duration 347115
+launch_time_start 1753933170729
+launch_time_stop 1753964196763
+launch_retries_retries 15
+launch_retries_run 48
