+++ allure-report/history/duration-trend.json
@@ -1,1 +1,1 @@
-[{"data":{"duration":118783}}]
+[{"data":{"duration":31026034}}]
