+++ allure-report/history/history-trend.json
@@ -1,1 +1,1 @@
-[{"data":{"failed":4,"broken":1,"skipped":1,"passed":43,"unknown":0,"total":49}}]
+[{"data":{"failed":0,"broken":0,"skipped":1,"passed":47,"unknown":0,"total":48}}]
