+++ allure-report/index.html
@@ -26,7 +26,7 @@
         gtag('js', new Date());
         gtag('config', 'G-FVWC4GKEYS', {
           'allureVersion': '2.34.0',
-          'reportUuid': 'f41eec71-49d1-4760-b117-a1ed7cdde8fd',
+          'reportUuid': 'da8b34aa-add2-476a-8ee8-cf6d20a9cf38',
           'single_file': false
         });
     </script>
