+++ allure-report/widgets/behaviors.json
@@ -1,1 +1,1 @@
-{"total":49,"items":[]}
+{"total":48,"items":[]}
