+++ allure-report/widgets/categories-trend.json
@@ -1,1 +1,1 @@
-[{"data":{"Product defects":4,"Test defects":1}}]
+[{"data":{}}]
