+++ allure-report/widgets/duration-trend.json
@@ -1,1 +1,1 @@
-[{"data":{"duration":118783}}]
+[{"data":{"duration":31026034}}]
