+++ allure-report/widgets/retry-trend.json
@@ -1,1 +1,1 @@
-[{"data":{"run":49,"retry":75}}]
+[{"data":{"run":48,"retry":15}}]
