+++ allure-report/widgets/suites.json
@@ -1,1 +1,1 @@
-{"total":31,"items":[{"uid":"e6753d74d8085e446aaa8a64ce395f0d","name":"BasicsTypeScript\\TS23_PostCallAuthToken.spec.ts","statistic":{"failed":2,"broken":0,"skipped":0,"passed":0,"unknown":0,"total":2}},{"uid":"85c55a40df1926269ceab2ddcb97f44e","name":"BasicsTypeScript\\TS09_GGRegPageDropDown.spec.ts","statistic":{"failed":1,"broken":0,"skipped":0,"passed":0,"unknown":0,"total":1}},{"uid":"d4c2f63bc3a59bf3620720ac01b614d9","name":"test-1.spec.ts","statistic":{"failed":1,"broken":0,"skipped":0,"passed":0,"unknown":0,"total":1}},{"uid":"990d931a332a1d859ba8b39ec0dcc472","name":"BasicsTypeScript\\copilotUsage.spec.ts","statistic":{"failed":0,"broken":1,"skipped":0,"passed":0,"unknown":0,"total":1}},{"uid":"252b7aa4342b2483ea87057ebca8ff4a","name":"BasicsTypeScript\\TS22_PostCallDDT.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":5,"unknown":0,"total":5}},{"uid":"8395cac1a9089639eafc92d912fac892","name":"BasicsTypeScript\\TS19_hooksUsage.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":3,"unknown":0,"total":3}},{"uid":"b7cafcb4eee3454e627dfab284d51185","name":"DDT\\DDT01_TestDataPutInArray.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":3,"unknown":0,"total":3}},{"uid":"8483b55b6b88b57cfa2fad92c0dcffb0","name":"DDT\\Excellogin.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":3,"unknown":0,"total":3}},{"uid":"f4413010d32c4d6449a672d422f0425b","name":"DDT\\DDT02_GetDataFromJsonFile.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":3,"unknown":0,"total":3}},{"uid":"dbdd51f1108a64be10a0c5ecdc1c759a","name":"BasicsTypeScript\\TS03_arrayLearing.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":2,"unknown":0,"total":2}}]}
+{"total":30,"items":[{"uid":"252b7aa4342b2483ea87057ebca8ff4a","name":"BasicsTypeScript\\TS22_PostCallDDT.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":5,"unknown":0,"total":5}},{"uid":"8395cac1a9089639eafc92d912fac892","name":"BasicsTypeScript\\TS19_hooksUsage.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":3,"unknown":0,"total":3}},{"uid":"b7cafcb4eee3454e627dfab284d51185","name":"DDT\\DDT01_TestDataPutInArray.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":3,"unknown":0,"total":3}},{"uid":"f4413010d32c4d6449a672d422f0425b","name":"DDT\\DDT02_GetDataFromJsonFile.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":3,"unknown":0,"total":3}},{"uid":"8483b55b6b88b57cfa2fad92c0dcffb0","name":"DDT\\Excellogin.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":3,"unknown":0,"total":3}},{"uid":"5f787dab93dc3b004c688a92e1204a04","name":"BasicsTypeScript\\ClientAppPageObject.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":2,"unknown":0,"total":2}},{"uid":"dbdd51f1108a64be10a0c5ecdc1c759a","name":"BasicsTypeScript\\TS03_arrayLearing.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":2,"unknown":0,"total":2}},{"uid":"fb811b7e9ac483f03eb2aeee312d36f5","name":"BasicsTypeScript\\TS06_TupleUsage.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":2,"unknown":0,"total":2}},{"uid":"ebd639322b4a754395b0af6cfc0c9f77","name":"example.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":2,"unknown":0,"total":2}},{"uid":"69d5d376d265cdb64fbcbb58e406e532","name":"BasicsTypeScript\\TS12_TagUsage.spec.ts","statistic":{"failed":0,"broken":0,"skipped":0,"passed":2,"unknown":0,"total":2}}]}
