+++ allure-report/widgets/summary.json
@@ -1,1 +1,1 @@
-{"reportName":"Allure Report","testRuns":[],"statistic":{"failed":4,"broken":1,"skipped":1,"passed":43,"unknown":0,"total":49},"time":{"start":1752973871942,"stop":1752973990725,"duration":118783,"minDuration":1,"maxDuration":91922,"sumDuration":328536}}
+{"reportName":"Allure Report","testRuns":[],"statistic":{"failed":0,"broken":0,"skipped":1,"passed":47,"unknown":0,"total":48},"time":{"start":1753933170729,"stop":1753964196763,"duration":31026034,"minDuration":1,"maxDuration":67996,"sumDuration":347115}}
