--- allure-results/ccd657d4-d732-4928-a0b0-2fa544e9d0bd-result.json
@@ -1,1 +1,0 @@
-{"uuid":"ccd657d4-d732-4928-a0b0-2fa544e9d0bd","name":"Valid login should succeed","historyId":"bb0b0ac901a11d33861e8cd40e84c4b1:d41d8cd98f00b204e9800998ecf8427e","status":"failed","statusDetails":{"message":"Error: page.goto: net::ERR_NAME_NOT_RESOLVED at https://qa.saucedemo.com/\nCall log:\n  - navigating to \"https://qa.saucedemo.com/\", waiting until \"load\"\n","trace":"Error: page.goto: net::ERR_NAME_NOT_RESOLVED at https://qa.saucedemo.com/\nCall log:\n  - navigating to \"https://qa.saucedemo.com/\", waiting until \"load\"\n\n    at C:\\Users\\softw\\TypescriptWorkspace\\STL-TypeScriptProject-Krishna\\tests\\BasicsTypeScript\\TS19_ReadEnvFile.spec.ts:5:16"},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"name":"browserType.launch","start":1748738580486,"uuid":"6829af2b-989b-4806-ba24-2df7060b7d5e","stop":1748738581086}],"attachments":[],"parameters":[],"name":"fixture: browser","start":1748738580480,"uuid":"36aae956-e9c6-46bb-912b-b1dcb55ae1ef","stop":1748738581087},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"name":"browser.newContext","start":1748738581091,"uuid":"9517d104-d054-4732-b7a0-a9eba100bd27","stop":1748738581107}],"attachments":[],"parameters":[],"name":"fixture: context","start":1748738581088,"uuid":"0336e177-83be-4399-b794-18094f567fa0","stop":1748738581122},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"name":"browserContext.newPage","start":1748738581126,"uuid":"6ea8915f-0f2d-42e9-b36e-9144e7767064","stop":1748738582255}],"attachments":[],"parameters":[],"name":"fixture: page","start":1748738581123,"uuid":"d990be00-4a21-4c8f-9aa0-8fb73e36c196","stop":1748738582255}],"attachments":[],"parameters":[],"name":"Before Hooks","start":1748738580457,"uuid":"5a085f3d-b752-40c8-994b-731308a85008","stop":1748738582264},{"status":"failed","statusDetails":{"message":"Error: page.goto: net::ERR_NAME_NOT_RESOLVED at https://qa.saucedemo.com/\nCall log:\n  - navigating to \"https://qa.saucedemo.com/\", waiting until \"load\"\n","trace":"Error: page.goto: net::ERR_NAME_NOT_RESOLVED at https://qa.saucedemo.com/\nCall log:\n  - navigating to \"https://qa.saucedemo.com/\", waiting until \"load\"\n\n    at C:\\Users\\softw\\TypescriptWorkspace\\STL-TypeScriptProject-Krishna\\tests\\BasicsTypeScript\\TS19_ReadEnvFile.spec.ts:5:16"},"stage":"finished","steps":[],"attachments":[],"parameters":[],"start":1748738582342,"name":"page.goto(https://qa.saucedemo.com)","uuid":"828290f2-31ac-4e51-8098-37abae994832","stop":1748738582807},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"screenshot","source":"297feebe-18bb-4338-a75a-eaf6db9ed768-attachment.png","type":"image/png"}],"parameters":[],"start":1748738584241,"name":"screenshot","stop":1748738584241},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"video","source":"c97c7b54-1c29-4daa-86df-1d1e12b1672f-attachment.webm","type":"video/webm"}],"parameters":[],"start":1748738584245,"name":"video","stop":1748738584246},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"_error-context-0","source":"c17cf88a-05b1-4842-9e6d-f854a9ab10e4-attachment.md","type":"text/markdown"}],"parameters":[],"start":1748738584249,"name":"_error-context-0","stop":1748738584249},{"statusDetails":{},"stage":"finished","steps":[],"attachments":[{"name":"trace","source":"599cf664-0345-4b33-875e-e072cc2db581-attachment.zip","type":"application/vnd.allure.playwright-trace"}],"parameters":[],"start":1748738584251,"name":"trace","stop":1748738584251},{"status":"passed","statusDetails":{},"stage":"finished","steps":[{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"name":"page.screenshot","start":1748738582822,"uuid":"cb306cb1-8be2-4f70-a1d4-3d32732a4d1a","stop":1748738583241},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"name":"fixture: page","start":1748738583248,"uuid":"cb8cd9d5-838f-4918-9d91-d2a3806ad299","stop":1748738583251},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"name":"fixture: context","start":1748738583252,"uuid":"2d8269e1-c62d-4979-87fc-a5343d49d212","stop":1748738583252},{"status":"passed","statusDetails":{},"stage":"finished","steps":[],"attachments":[],"parameters":[],"name":"video.saveAs","start":1748738583607,"uuid":"76c617b9-5c3b-4aa8-9219-a01ec1eb046c","stop":1748738583610}],"attachments":[],"parameters":[],"name":"After Hooks","start":1748738582814,"uuid":"4e1270e7-c88c-404f-947e-c54d1ac7bdc6","stop":1748738583621}],"attachments":[],"parameters":[],"labels":[{"name":"language","value":"javascript"},{"name":"framework","value":"playwright"},{"name":"package","value":"BasicsTypeScript.TS19_ReadEnvFile.spec.ts"},{"name":"titlePath","value":" >  > BasicsTypeScript\\TS19_ReadEnvFile.spec.ts"},{"name":"host","value":"LAPTOP-PRVS3R3O"},{"name":"thread","value":"pid-24000-worker-0"},{"name":"suite","value":"BasicsTypeScript\\TS19_ReadEnvFile.spec.ts"}],"links":[],"start":1748738580457,"testCaseId":"bb0b0ac901a11d33861e8cd40e84c4b1","fullName":"BasicsTypeScript/TS19_ReadEnvFile.spec.ts:4:7","stop":1748738582971}
