+++ package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "playwright test",
+    "clean:allure": "rmdir /s /q allure-results && rmdir /s /q allure-report",
     "allure:generate": "allure generate allure-results --clean -o allure-report",
     "allure:open": "allure open allure-report"
   },
@@ -14,17 +15,19 @@
   "devDependencies": {
     "@playwright/test": "^1.54.1",
     "@types/node": "^22.16.4",
-    "allure-commandline": "^2.34.0",
-    "allure-playwright": "^3.2.2",
+    "allure-commandline": "^2.34.1",
+    "allure-playwright": "^3.3.2",
     "cross-env": "^7.0.3",
     "eslint": "^9.31.0",
     "eslint-config-prettier": "^10.1.5",
     "eslint-plugin-playwright": "^2.2.0",
+    "fs-extra": "^11.3.0",
     "prettier": "^3.6.2",
     "typescript": "^5.8.3",
     "xlsx": "^0.18.5"
   },
   "dependencies": {
-    "dotenv": "^16.5.0"
+    "dotenv": "^16.5.0",
+    "latest": "^0.2.0"
   }
 }
