--- tests/test-1.spec.ts
@@ -1,8 +1,0 @@
-import { test, expect } from '@playwright/test';
-
-test('test', async ({ page }) => {
-  await page.goto('https://www.thegoodguys.com.au/');
-  await page.getByRole('link', { name: '' }).click();
-  await page.getByRole('button', { name: 'CREATE ACCOUNT' }).click();
-  await page.getByRole('button', { name: 'CREATE ACCOUNT' }).click();
-});
