+++ utils/placeorderTestData.json
@@ -5,8 +5,8 @@
     "item": "IPHONE 13 PRO"
 },
 {
-    "username" : "anshika@gmail.com",
-    "password" : "Iamking@000",
+    "username" : "hinatesting30@gmail.com",
+    "password" : "@Aisha1983",
     "item": "IPHONE 13 PRO"
 }
 ]
